+++ SwiftUI_Pull_to_Refresh/Views/CitiesView.swift
@@ -10,84 +10,53 @@ import Introspect
 import OpenWeather
 
 
+private struct CitiesFrameEnvironmentKey: EnvironmentKey {
+	static let defaultValue: CGRect = .zero
+}
+
+extension EnvironmentValues {
+	var citiesFrame: CGRect {
+		get { self[CitiesFrameEnvironmentKey.self] }
+		set { self[CitiesFrameEnvironmentKey.self] = newValue }
+	}
+}
+
+
 struct CitiesView: View {
+	
+	let viewModel = CitiesViewModel()
 	
 	var body: some View {
 		ZStack {
-			Color.clear
+			
+			// Background.
+			Color.orange.opacity(0.2)
 				.overlay(
 					VStack {
-						Image("WorldMap")
-							.opacity(0.2)
+						UI.Image.worldMap
+							.opacity(0.5)
 						Spacer()
 					}
+						
 				)
+			
+			// Cities.
 			GeometryReader { geometry in
 				ScrollView(.horizontal, showsIndicators: false) {
 					HStack(spacing: 0) {
-						CityView(
-							viewModel: CityViewModel(
-								name: "San Francisco",
-								location: .init(latitude: 37.7749, longitude: 122.4194)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "New York",
-								location: .init(latitude: 40.7128, longitude: 74.0060)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "Paris",
-								location: .init(latitude: 48.8566, longitude: 2.3522)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "London",
-								location: .init(latitude: 51.5074, longitude: 0.1278)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "Moscow",
-								location: .init(latitude: 55.7558, longitude: 37.6173)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "New Delhi",
-								location: .init(latitude: 28.6139, longitude: 77.2090)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "Tokyo",
-								location: .init(latitude: 35.6762, longitude: 139.6503)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "Sidney",
-								location: .init(latitude: 33.8688, longitude: 151.2093)
-							),
-							width: geometry.size.width
-						)
-						CityView(
-							viewModel: CityViewModel(
-								name: "Honolulu",
-								location: .init(latitude: 21.3069, longitude: 157.8583)
-							),
-							width: geometry.size.width
-						)
+						ForEach(viewModel.cities) { eachCityViewModel in
+							CityView(
+								viewModel: CityViewModel(
+									name: eachCityViewModel.name,
+									location: .init(
+										latitude: eachCityViewModel.location.latitude,
+										longitude: eachCityViewModel.location.longitude
+									)
+								),
+								width: geometry.size.width
+							)
+								.environment(\.citiesFrame,  geometry.frame(in: .global))
+						}
 					}
 				}
 				.introspectScrollView {